02 clase fastapi/pyproject.toml
[project]
name = "02-clase-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "pandas>=3.0.2",
    "psycopg[binary,pool]>=3.3.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.45.0",
]
